% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Title Brute Force-Knapsack}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data frame consist of two columns w and v}

\item{W}{total capacity}

\item{parallel}{To Support for parallel computation}
}
\value{
Algorithm that goes through all possible alternatives and return the maximum value found
}
\description{
Title Brute Force-Knapsack
}
\examples{
RNGversion(min(as.character(getRversion()),"3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <- data.frame(
w=sample(1:4000, size = n, replace = TRUE),
 v=runif(n = n, 0, 10000))
 brute_force_knapsack(x = knapsack_objects[1:8,], W = 3500)
}
