% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_knapsack.R
\name{dynamic_knapsack}
\alias{dynamic_knapsack}
\title{Title Dynamic Programming-Knapsack}
\usage{
dynamic_knapsack(x, W)
}
\arguments{
\item{x}{Data frame consist of two columns w and v}

\item{W}{Total Capacity}
}
\value{
Algorithm that can solve the knapsack problem exact by iterating over all possible values of w.
}
\description{
Title Dynamic Programming-Knapsack
}
\examples{
RNGversion(min(as.character(getRversion()),"3.5.3"))
set.seed(42, kind = "Mersenne-Twister", normal.kind = "Inversion")
n <- 2000
knapsack_objects <- data.frame(
w=sample(1:4000, size = n, replace = TRUE),
 v=runif(n = n, 0, 10000))
 dynamic_knapsack(x = knapsack_objects[1:8,], W = 3500)
}
